MyShop-Project
This project is a frontend design of an e-commerce website built with React. Global state management is handled using Redux Toolkit, and data is fetched from APIs using Axios. A modern user interface is designed with Material UI (MUI). Product data is retrieved from fakestoreapi.com, and comments are fetched from jsonplaceholder.typicode.com. The project focuses on the use of Redux slices and asynchronous programming.